# openapi/generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DataStack API",
    "version": "2.0.0",
    "description": "REST API for DataStack: authentication, workspaces, compute clusters, jobs, notebooks, pipelines, SQL warehouses, and webhooks."
  },
...
  "paths": {
...
    "/v1/clusters": {
      "get": {
        "summary": "List clusters",
        "tags": [
          "Compute / Clusters"
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "RUNNING",
                "RESIZING",
                "TERMINATING",
                "TERMINATED",
                "ERROR"
              ]
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 25
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "clusters": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "workspaceId": {
                            "type": "string"
                          },
                          "clusterSize": {
                            "type": "string"
                          },
                          "region": {
                            "type": "string",
                            "description": "Cloud region where the cluster runs (e.g. us-east-1)"
                          },
                          "state": {
                            "type": "string",
                            "enum": [
                              "PENDING",
                              "RUNNING",
                              "RESIZING",
                              "TERMINATING",
                              "TERMINATED",
                              "ERROR"
                            ]
                          },
                          "sparkVersion": {
                            "type": "string"
                          },
                          "nodeType": {
                            "type": "string"
                          },
                          "workerCount": {
                            "type": "integer"
                          },
                          "minWorkers": {
                            "type": "integer"
                          },
                          "maxWorkers": {
                            "type": "integer"
                          },
                          "enableAutoscaling": {
                            "type": "boolean"
                          },
                          "autoTerminationMinutes": {
                            "type": "integer"
                          },
                          "tags": {
                            "type": "object",
                            "additionalProperties": {
                              "type": "string"
                            }
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "createdBy": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "workspaceId",
                          "region",
                          "state",
                          "createdAt"
                        ]
                      }
                    },
                    "totalCount": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "clusters",
                    "totalCount"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create cluster",
        "tags": [
          "Compute / Clusters"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "workspaceId": {
                    "type": "string"
                  },
                  "sparkVersion": {
                    "type": "string"
                  },
                  "nodeType": {
                    "type": "string"
                  },
                  "workerCount": {
                    "type": "integer"
                  },
                  "enableAutoscaling": {
                    "type": "boolean"
                  },
                  "minWorkers": {
                    "type": "integer"
                  },
                  "maxWorkers": {
                    "type": "integer"
                  },
                  "autoTerminationMinutes": {
                    "type": "integer"
                  },
                  "tags": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "name",
                  "workspaceId",
                  "sparkVersion",
                  "nodeType"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "workspaceId": {
                      "type": "string"
                    },
                    "clusterSize": {
                      "type": "string"
                    },
                    "region": {
                      "type": "string",
                      "description": "Cloud region where the cluster runs (e.g. us-east-1)"
                    },
                    "state": {
                      "type": "string",
                      "enum": [
                        "PENDING",
                        "RUNNING",
                        "RESIZING",
                        "TERMINATING",
                        "TERMINATED",
                        "ERROR"
                      ]
                    },
                    "sparkVersion": {
                      "type": "string"
                    },
                    "nodeType": {
                      "type": "string"
                    },
                    "workerCount": {
                      "type": "integer"
                    },
                    "minWorkers": {
                      "type": "integer"
                    },
                    "maxWorkers": {
                      "type": "integer"
                    },
                    "enableAutoscaling": {
                      "type": "boolean"
                    },
                    "autoTerminationMinutes": {
                      "type": "integer"
                    },
                    "tags": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "createdBy": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "workspaceId",
                    "region",
                    "state",
                    "createdAt"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/clusters/{clusterId}": {
      "get": {
        "summary": "Get cluster by ID",
        "tags": [
          "Compute / Clusters"
        ],
        "parameters": [
          {
            "name": "clusterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "workspaceId": {
                      "type": "string"
                    },
                    "clusterSize": {
                      "type": "string"
                    },
                    "region": {
                      "type": "string",
                      "description": "Cloud region where the cluster runs (e.g. us-east-1)"
                    },
                    "state": {
                      "type": "string",
                      "enum": [
                        "PENDING",
                        "RUNNING",
                        "RESIZING",
                        "TERMINATING",
                        "TERMINATED",
                        "ERROR"
                      ]
                    },
                    "sparkVersion": {
                      "type": "string"
                    },
                    "nodeType": {
                      "type": "string"
                    },
                    "workerCount": {
                      "type": "integer"
                    },
                    "minWorkers": {
                      "type": "integer"
                    },
                    "maxWorkers": {
                      "type": "integer"
                    },
                    "enableAutoscaling": {
                      "type": "boolean"
                    },
                    "autoTerminationMinutes": {
                      "type": "integer"
                    },
                    "tags": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "createdBy": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "workspaceId",
                    "region",
                    "state",
                    "createdAt"
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Terminate cluster",
        "tags": [
          "Compute / Clusters"
        ],
        "parameters": [
          {
            "name": "clusterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/clusters/{clusterId}/resize": {
      "post": {
        "summary": "Resize cluster",
        "tags": [
          "Compute / Clusters"
        ],
        "parameters": [
          {
            "name": "clusterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "workerCount": {
                    "type": "integer"
                  },
                  "minWorkers": {
                    "type": "integer"
                  },
                  "maxWorkers": {
                    "type": "integer"
                  }
                },
                "required": [
                  "workerCount"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "workspaceId": {
                      "type": "string"
                    },
                    "clusterSize": {
                      "type": "string"
                    },
                    "region": {
                      "type": "string",
                      "description": "Cloud region where the cluster runs (e.g. us-east-1)"
                    },
                    "state": {
                      "type": "string",
                      "enum": [
                        "PENDING",
                        "RUNNING",
                        "RESIZING",
                        "TERMINATING",
                        "TERMINATED",
                        "ERROR"
                      ]
                    },
                    "sparkVersion": {
                      "type": "string"
                    },
                    "nodeType": {
                      "type": "string"
                    },
                    "workerCount": {
                      "type": "integer"
                    },
                    "minWorkers": {
                      "type": "integer"
                    },
                    "maxWorkers": {
                      "type": "integer"
                    },
                    "enableAutoscaling": {
                      "type": "boolean"
                    },
                    "autoTerminationMinutes": {
                      "type": "integer"
                    },
                    "tags": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "createdBy": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "workspaceId",
                    "region",
                    "state",
                    "createdAt"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}